% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chapters.R
\name{add_chapter}
\alias{add_chapter}
\title{Add a new chapter file to your course}
\usage{
add_chapter(chapter_name)
}
\arguments{
\item{chapter_name}{- the new chapter name, such as "chapter2.md"}
}
\value{
the new chapter file with YAML, which is saved in `chapters/`. RStudio
will open the file for editing.
}
\description{
Add a new chapter file to your course
}
