% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{add_data_to_lesson}
\alias{add_data_to_lesson}
\title{Adds a data object to both the master and binder branch}
\usage{
add_data_to_lesson(dataobj)
}
\arguments{
\item{dataobj}{- a data object, usually a tibble or data.frame}
}
\value{
added and committed objects to both the binder and master branches
}
\description{
Because the course repo uses both a master and binder branch, we need
to make sure the data objects are going in both branches.
}
\details{
The function will apply git stash to the repository before doing anything.
Then the object is added to both the master and binder branch using
usethis::use_data() before being committed to both.

The function then switches back to the current repository and
applies the stash to get back any uncommitted changes.
}
