% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chapters.R
\name{add_exercise}
\alias{add_exercise}
\title{Adds an exercise block to the end of a chapter}
\usage{
add_exercise(chapter_file = "chapter1.md", exercise_id = "01_01")
}
\arguments{
\item{chapter_file}{- name of the chapter, such as "chapter1.md"
without the path}

\item{exercise_id}{- id of the exercise, such as "01_03"}
}
\value{
- opened files including the chapter file, with the new
exercise/codeblock appended and the exercises opened
}
\description{
Adds an exercise block to the end of a chapter
}
