% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_lesson_repo}
\alias{create_lesson_repo}
\title{Create a new lesson repo}
\usage{
create_lesson_repo(destdir = NULL)
}
\arguments{
\item{destdir}{- destination directory for creating the course repository.
It will be created as `course-starter-r`.}
}
\value{
new project as created by usethis::create_from_github
}
\description{
Use this function to start a brand new course based on the template available
at https://github.com/ines/course-starter-r/
}
