% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chapters.R
\name{get_exercises}
\alias{get_exercises}
\title{Title}
\usage{
get_exercises(chapter_file)
}
\arguments{
\item{chapter_file}{}
}
\value{

}
\description{
Title
}
\examples{
chapter_file_path <- system.file("extdata/", package="decampr")
chapter_list <- get_chapters(chapter_file_path)
exercise_list <- get_exercises(chapter_list[[1]])
exercise_list[[1]]
}
