% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chapters.R
\name{make_yaml_block}
\alias{make_yaml_block}
\title{Read and write an appropriate YAML block}
\usage{
make_yaml_block(chapter_name, chapter_file_path)
}
\arguments{
\item{chapter_name}{- chapter name without full path, such as "chapter1.md"}

\item{chapter_file_path}{- full file path to the chapter.md file in the
DataCamp repository, such as "c:/Code/RBootcamp_old/". Set this argument
to NULL if you don't have a chapter heading.}
}
\value{
yaml block as a glue class
}
\description{
Read and write an appropriate YAML block
}
