% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chapters.R
\name{open_chapter}
\alias{open_chapter}
\title{Opens a chapter from the file name}
\usage{
open_chapter(chapter_name)
}
\arguments{
\item{chapter_name}{- name of the chapter to open. Note that chapter names
are numbered numerically (examples: 'chapter1.md', 'chapter2.md').}
}
\value{
opened chapter file if in an interactive session
}
\description{
Opens a chapter from the file name
}
