% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chapters.R
\name{open_exercise}
\alias{open_exercise}
\title{Opens exercise files for editing}
\usage{
open_exercise(id, create = FALSE)
}
\arguments{
\item{id}{- id of the exercise/solution/pre-exercise (such as "01_02")}

\item{create}{- boolean. If FALSE, won't create the relevant files and returns NULL.
If TRUE, then will create files in exercises directory}
}
\value{
opens exercise files
}
\description{
Opens exercise files for editing
}
