% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chapters.R
\name{save_exercise_list}
\alias{save_exercise_list}
\title{Given an exercise list and a chapter name, writes files.}
\usage{
save_exercise_list(ex_list, chapter_name)
}
\arguments{
\item{chapter_name}{}
}
\value{
written exercises/solutions in `exercises/` and written chapter in `chapters/`
}
\description{
Given an exercise list and a chapter name, writes files.
}
